requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-decouple==3.8
requests==2.32.3
schedule==1.2.2
termcolor==2.5.0
urllib3==2.2.3
